First commit
This code is working, but it's not very well organized, it could be better.
The server must be run first, and then the client. When the client is run, if it doesn't have the required sound files, it will ask them of the server, and download them. Then it will behave as a messaging client.